+++ src/Server/Features/Admin/RefreshPlayerHeadshots.cs
@@ -1,11 +1,11 @@
-﻿using DynamoLeagueBlazor.Server.Infrastructure;
+﻿using DynamoLeagueBlazor.Server.Features.Admin.Shared;
+using DynamoLeagueBlazor.Server.Infrastructure;
 using DynamoLeagueBlazor.Shared.Features.Admin;
 using DynamoLeagueBlazor.Shared.Infastructure.Identity;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System.Text.Json.Serialization;
 
 namespace DynamoLeagueBlazor.Server.Features.Admin;
 
@@ -34,7 +34,6 @@ public record RefreshPlayerHeadshotsCommand : IRequest<Unit> { }
 
 public class RefreshPlayerHeadshotsHandler : IRequestHandler<RefreshPlayerHeadshotsCommand>
 {
-    public const string PlayerProfilerUri = "https://www.playerprofiler.com/api/v1";
     private readonly ApplicationDbContext _dbContext;
     private readonly HttpClient _httpClient;
 
@@ -46,8 +45,7 @@ public class RefreshPlayerHeadshotsHandler : IRequestHandler<RefreshPlayerHeadsh
 
     public async Task<Unit> Handle(RefreshPlayerHeadshotsCommand request, CancellationToken cancellationToken)
     {
-        var playersUri = $"{PlayerProfilerUri}/players";
-        var playerResult = await _httpClient.GetFromJsonAsync<PlayerDataResult>(playersUri, cancellationToken: cancellationToken);
+        var playerResult = await _httpClient.GetFromJsonAsync<PlayerDataResult>(PlayerProfilerRouteFactory.GetPlayersUri, cancellationToken: cancellationToken);
 
         if (playerResult != null)
         {
@@ -62,8 +60,7 @@ public class RefreshPlayerHeadshotsHandler : IRequestHandler<RefreshPlayerHeadsh
 
                 if (matchingDynamoLeaguePlayer is null) continue;
 
-                var playerUri = $"{PlayerProfilerUri}/player/{player.PlayerId}";
-                var metricResult = await _httpClient.GetFromJsonAsync<PlayerMetricDataResult>(playerUri, cancellationToken: cancellationToken);
+                var metricResult = await _httpClient.GetFromJsonAsync<PlayerMetricDataResult>(PlayerProfilerRouteFactory.GetPlayerUri(player.PlayerId), cancellationToken: cancellationToken);
 
                 matchingDynamoLeaguePlayer.HeadShotUrl = metricResult!.Data.Player.Core.Avatar;
             }
@@ -72,48 +69,5 @@ public class RefreshPlayerHeadshotsHandler : IRequestHandler<RefreshPlayerHeadsh
         await _dbContext.SaveChangesAsync(cancellationToken);
 
         return Unit.Value;
-    }
-
-    internal class PlayerDataResult
-    {
-        public PlayerListResult Data { get; set; } = null!;
-
-        public class PlayerListResult
-        {
-            public IEnumerable<PlayerResult> Players { get; set; } = Array.Empty<PlayerResult>();
-
-            public class PlayerResult
-            {
-                [JsonPropertyName("Player_ID")]
-                public string PlayerId { get; set; } = null!;
-                [JsonPropertyName("Full Name")]
-                public string FullName { get; set; } = null!;
-                public string Team { get; set; } = null!;
-                public string Position { get; set; } = null!;
-            }
-        }
-    }
-
-    internal class PlayerMetricDataResult
-    {
-        public PlayerMetricData Data { get; set; } = null!;
-
-        public class PlayerMetricData
-        {
-            public PlayerData Player { get; set; } = null!;
-
-            public class PlayerData
-            {
-                [JsonPropertyName("Player_ID")]
-                public string PlayerId { get; set; } = null!;
-
-                public Player Core { get; set; } = null!;
-
-                public class Player
-                {
-                    public string Avatar { get; set; } = null!;
-                }
-            }
-        }
     }
 }
